+++ src/app/App.tsx
@@ -1,6 +1,6 @@
-import Router from './Router';
 import QueryProvider from './queryProvider';
-import { SocketProvider } from './socketProvider'; // 방금 만든 파일 import
+import Router from './Router';
+import { SocketProvider } from './socketProvider';
 
 const App = () => {
   return (
