+++ src/app/queryProvider.tsx
@@ -1,9 +1,11 @@
+import type { ReactNode } from 'react';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { queryClient } from '@/shared/api/queryClient';
+
+import { queryClient } from '@/shared/api';
 
 interface QueryProviderProps {
-  children: React.ReactNode;
+  children: ReactNode;
 }
 
 const QueryProvider = ({ children }: QueryProviderProps) => {
