+++ src/entities/chat/ui/MentionDropdown.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 import type { Friend } from '@/shared/model';
 
 interface MentionDropdownProps {
