+++ src/entities/game/ui/ItemIcon.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 export interface ItemIconProps {
   id: string;
