+++ src/entities/room/api/createRoom.ts
@@ -1,4 +1,4 @@
-import { instance } from '@/shared';
+import { instance } from '@/shared/api';
 import type { CreateRoomPayload, Room } from '../model/types';
 
 export const createRoom = async (data: CreateRoomPayload): Promise<Room> => {
