+++ src/features/auth/api/postRefreshToken.ts
@@ -1,4 +1,4 @@
-import { instance } from '@/shared/api/axios';
+import { instance } from '@/shared/api';
 
 export const postRefreshToken = async () => {
   const response = await instance.post('auth/refresh');
