+++ src/features/lobby/ui/MenuButton.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react';
 
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 import { COLOR_MAP } from '../constants/colors';
 
 interface MenuButtonProps {
