+++ src/features/lobby/ui/RoomActionButtons.tsx
@@ -1,6 +1,6 @@
 import type { ButtonHTMLAttributes } from 'react';
 
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 interface RoomActionButtonsProps {
   amIHost: boolean;
