+++ src/features/main/ui/CreateRoomModal/VisibilitySelector.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 interface VisibilitySelectorProps {
   visibility: 'public' | 'private' | null;
