+++ src/features/main/ui/RoomList/RoomCard.tsx
@@ -5,7 +5,7 @@ import {
 
 import type { Room } from '@/entities/room';
 import { ROOM_MODE, ROOM_STATUS } from '@/entities/room';
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 interface RoomCardProps {
   room: Room;
